backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic[email]
python-dotenv
passlib[bcrypt]
python-jose[cryptography]
celery
redis